third_party/blink/web_tests/virtual/xslt-disabled/fast/xsl/xslt-bad-import-uri-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been removed in this browser. See https://chromestatus.com/feature/4709671889534976.
This tests that a bad @import URI does not cause a crash. 
